hw_13/requirements.txt
# Requirements for Jupyter Notebook (Google Colab)
# Most packages are pre-installed on Colab, but listed here for local setup

transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.0
trl>=0.7.4
bitsandbytes>=0.41.0
accelerate>=0.25.0
beautifulsoup4>=4.12.0
requests>=2.31.0
torch>=2.1.0
python-dotenv>=1.0.0

